# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Customer API",
        "description": "API documentation for customer registration and authentication",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/service/store": {
            "post": {
                "tags": [
                    "Services"
                ],
                "summary": "Create a new service for a customer",
                "operationId": "a2f022af0779d76752a203556374a2a6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "customer_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Internet Plan"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Unlimited 100 Mbps"
                                    },
                                    "customer_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Service created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ServiceResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/service/customer/{customer}": {
            "get": {
                "tags": [
                    "Services"
                ],
                "summary": "Get services of a customer",
                "operationId": "699d41c3b14a15edcd9ba09995346ce3",
                "parameters": [
                    {
                        "name": "customer",
                        "in": "path",
                        "description": "Customer ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of services for a customer",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ServicesOfCustomer"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Customer not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/service/all": {
            "get": {
                "tags": [
                    "Services"
                ],
                "summary": "Get all services with customers",
                "operationId": "ea96368be51e6a69e50719963ffa164b",
                "responses": {
                    "200": {
                        "description": "Paginated list of all services",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/AllServicesResource"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/service/update/{service}": {
            "patch": {
                "tags": [
                    "Services"
                ],
                "summary": "Update an existing service",
                "operationId": "1f273c4e5e57264954169198c600bd67",
                "parameters": [
                    {
                        "name": "service",
                        "in": "path",
                        "description": "Service ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Plan"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated description"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Service updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ServiceResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Service not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/service/delete/{service}": {
            "delete": {
                "tags": [
                    "Services"
                ],
                "summary": "Delete a service",
                "operationId": "b5c898e7ba0c1fbbe5edf110744619df",
                "parameters": [
                    {
                        "name": "service",
                        "in": "path",
                        "description": "Service ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Service deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Service successfully deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Service not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "AllServicesResource": {
                "title": "AllServicesResource",
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Hosting Plan"
                    },
                    "description": {
                        "type": "string",
                        "example": "Basic web hosting"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-07-29T10:30:00Z"
                    }
                },
                "type": "object"
            },
            "ServiceResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Internet Plan"
                    },
                    "description": {
                        "type": "string",
                        "example": "Unlimited 100 Mbps"
                    },
                    "customer_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "ServicesOfCustomer": {
                "title": "Services of Customer",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Internet Package"
                    },
                    "description": {
                        "type": "string",
                        "example": "High speed internet"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 29.99
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}